Refactor: Extract logic from large activity and repository classes
- Moved UI setup, preferences, and section display logic from `DashboardActivity` into dedicated `DashboardActivityPreferences` and `DashboardActivitySections` files.
- Extracted team member UI adapters and state management into `DashboardTeamMembersAdapters` and `DashboardTeamMembersStateSupport`.
- Delegated team fetching and management operations from `TeamsFragment` to `TeamsFragmentOperations`.
- Relocated `OpenAiTranslationClient` and audio waveform extraction to their own utility files.
- Split `DashboardCoursesRepository` and `DashboardTeamsOperations` implementation details into separate operation files for better modularity.

This improves code organization, readability, and maintainability by adhering to the single responsibility principle.
